requirements.txt
Flask
azure-cli